# pkg/apispec/openapi.yaml
openapi: 3.0.3
info:
  title: Sandbox0 API
  version: 0.1.0
  description: Public HTTP APIs exposed by Sandbox0 regional gateways and the global gateway.
servers:
  - url: https://api.sandbox0.ai
    description: Production
...
paths:
...
  /api/v1/sandboxes:
    get:
      tags: [sandboxes]
      summary: List sandboxes
      description: List all sandboxes for the authenticated team. In multi-cluster mode, this endpoint aggregates results from all enabled clusters.
      security:
        - bearerAuth: []
      x-upstream-service: scheduler
      x-upstream-path: /api/v1/sandboxes
      parameters:
        - name: status
          in: query
          description: Filter by sandbox status
          schema:
            type: string
            enum: [starting, running, failed, completed]
        - name: template_id
          in: query
          description: Filter by template ID
          schema:
            type: string
        - name: paused
          in: query
          description: Filter by paused state
          schema:
            type: boolean
        - name: limit
          in: query
          description: Maximum number of results per page
          schema:
            type: integer
            default: 50
            minimum: 1
            maximum: 200
        - name: offset
          in: query
          description: Pagination offset
          schema:
            type: integer
            default: 0
            minimum: 0
      responses:
        "200":
          description: List of sandboxes
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessSandboxListResponse"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
    post:
      tags: [sandboxes]
      summary: Create (claim) a sandbox
      security:
        - bearerAuth: []
      x-upstream-service: manager
      x-upstream-path: /api/v1/sandboxes
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClaimRequest"
      responses:
        "201":
          description: Sandbox claimed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessClaimResponse"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /api/v1/sandboxes/{id}:
    get:
      tags: [sandboxes]
      summary: Get sandbox by ID
      security:
        - bearerAuth: []
      x-upstream-service: manager
      x-upstream-path: /api/v1/sandboxes/{id}
      parameters:
        - $ref: "#/components/parameters/SandboxID"
      responses:
        "200":
          description: Sandbox
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessSandboxResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
    put:
      tags: [sandboxes]
      summary: Update sandbox configuration
      security:
        - bearerAuth: []
      x-upstream-service: manager
      x-upstream-path: /api/v1/sandboxes/{id}
      parameters:
        - $ref: "#/components/parameters/SandboxID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SandboxUpdateRequest"
      responses:
        "200":
          description: Sandbox updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessSandboxResponse"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
    delete:
      tags: [sandboxes]
      summary: Delete (terminate) a sandbox
      security:
        - bearerAuth: []
      x-upstream-service: manager
      x-upstream-path: /api/v1/sandboxes/{id}
      parameters:
        - $ref: "#/components/parameters/SandboxID"
      responses:
        "200":
          description: Sandbox terminated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessMessageResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /api/v1/sandboxes/{id}/status:
    get:
      tags: [sandboxes]
      summary: Get sandbox status
      security:
        - bearerAuth: []
      x-upstream-service: manager
      x-upstream-path: /api/v1/sandboxes/{id}/status
      parameters:
        - $ref: "#/components/parameters/SandboxID"
      responses:
        "200":
          description: Sandbox status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessSandboxStatusResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /api/v1/sandboxes/{id}/logs:
    get:
      tags: [sandboxes]
      summary: Get sandbox process logs
      description: |
        Returns sandbox process output mirrored through the sandbox main container.
        Procd service logs are filtered out and remain available through Kubernetes pod logs.
        When `follow=false`, the response is a bounded text/plain snapshot.
        When `follow=true`, the response is a text/plain stream until the client disconnects.
        Kubernetes log selection parameters such as `tail_lines` and `limit_bytes`
        are applied before procd service log filtering.
      security:
        - bearerAuth: []
      x-upstream-service: manager
      x-upstream-path: /api/v1/sandboxes/{id}/logs
      parameters:
        - $ref: "#/components/parameters/SandboxID"
        - name: container
          in: query
          description: Pod container name. Defaults to the sandbox main container.
          schema:
            type: string
            default: procd
        - name: tail_lines
          in: query
          description: Maximum number of Kubernetes log lines read from the end of the log before procd service log filtering.
          schema:
            type: integer
            format: int64
            default: 200
            minimum: 1
            maximum: 5000
        - name: limit_bytes
          in: query
          description: Maximum response log payload bytes read from Kubernetes. Defaults only apply when follow is false.
          schema:
            type: integer
            format: int64
            default: 1048576
            minimum: 1
            maximum: 8388608
        - name: follow
          in: query
          description: Stream logs as text/plain until the client disconnects. When false, return a text/plain snapshot.
          schema:
            type: boolean
            default: false
        - name: previous
          in: query
          description: Return logs for the previously terminated container instance.
          schema:
            type: boolean
            default: false
        - name: timestamps
          in: query
          description: Include Kubernetes log timestamps when available.
          schema:
            type: boolean
            default: false
        - name: since_seconds
          in: query
          description: Only return logs newer than this many seconds.
          schema:
            type: integer
            format: int64
            minimum: 1
      responses:
        "200":
          description: Sandbox process logs
          headers:
            X-Sandbox-ID:
              description: Sandbox ID for the returned log payload.
              schema:
                type: string
            X-Sandbox-Pod-Name:
              description: Kubernetes pod name that produced the returned log payload.
              schema:
                type: string
            X-Sandbox-Log-Container:
              description: Kubernetes container name that produced the returned log payload.
              schema:
                type: string
            X-Sandbox-Log-Previous:
              description: Whether the returned payload came from a previously terminated container instance.
              schema:
                type: boolean
          content:
            text/plain:
              schema:
                type: string
        "400":
          description: Invalid log query parameters
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /api/v1/sandboxes/{id}/pause:
    post:
      tags: [sandboxes]
      summary: Pause a sandbox
      security:
        - bearerAuth: []
      x-upstream-service: manager
      x-upstream-path: /api/v1/sandboxes/{id}/pause
      parameters:
        - $ref: "#/components/parameters/SandboxID"
      responses:
        "200":
          description: Sandbox paused
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessPauseSandboxResponse"
        "409":
          description: Pause was superseded by a newer power transition
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "504":
          description: Timed out waiting for the sandbox to pause
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /api/v1/sandboxes/{id}/resume:
    post:
      tags: [sandboxes]
      summary: Resume a sandbox
      security:
        - bearerAuth: []
      x-upstream-service: manager
      x-upstream-path: /api/v1/sandboxes/{id}/resume
      parameters:
        - $ref: "#/components/parameters/SandboxID"
      responses:
        "200":
          description: Sandbox resumed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResumeSandboxResponse"
        "409":
          description: Resume was superseded by a newer power transition
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "504":
          description: Timed out waiting for the sandbox to resume
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /api/v1/sandboxes/{id}/refresh:
    post:
      tags: [sandboxes]
      summary: Refresh sandbox TTL
      security:
        - bearerAuth: []
      x-upstream-service: manager
      x-upstream-path: /api/v1/sandboxes/{id}/refresh
      parameters:
        - $ref: "#/components/parameters/SandboxID"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SandboxRefreshRequest"
      responses:
        "200":
          description: TTL refreshed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessRefreshResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /api/v1/sandboxes/{id}/network:
    get:
      tags: [sandboxes]
      summary: Get sandbox network policy
      security:
        - bearerAuth: []
      x-upstream-service: manager
      x-upstream-path: /api/v1/sandboxes/{id}/network
      parameters:
        - $ref: "#/components/parameters/SandboxID"
      responses:
        "200":
          description: Network policy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessSandboxNetworkPolicyResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
    put:
      tags: [sandboxes]
      summary: Update sandbox network policy
      security:
        - bearerAuth: []
      x-upstream-service: manager
      x-upstream-path: /api/v1/sandboxes/{id}/network
      parameters:
        - $ref: "#/components/parameters/SandboxID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SandboxNetworkPolicy"
      responses:
        "200":
          description: Network policy updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessSandboxNetworkPolicyResponse"
  /api/v1/sandboxes/{id}/public-gateway:
    get:
      tags: [sandboxes]
      summary: Get sandbox public gateway policy
      security:
        - bearerAuth: []
      x-upstream-service: manager
      x-upstream-path: /api/v1/sandboxes/{id}/public-gateway
      parameters:
        - $ref: "#/components/parameters/SandboxID"
      responses:
        "200":
          description: Public gateway policy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessPublicGatewayResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
    put:
      tags: [sandboxes]
      summary: Update sandbox public gateway policy
      description: |
        Replaces the request-level public gateway policy for sandbox public traffic.
        When enabled, matching routes control method, path, authentication, CORS,
        per-route rate limits, timeout, rewrite, and paused sandbox auto-resume.
      security:
        - bearerAuth: []
      x-upstream-service: manager
      x-upstream-path: /api/v1/sandboxes/{id}/public-gateway
      parameters:
        - $ref: "#/components/parameters/SandboxID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PublicGatewayConfig"
      responses:
        "200":
          description: Public gateway policy updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessPublicGatewayResponse"
        "400":
          description: Invalid public gateway policy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /api/v1/sandboxes/{id}/services:
    get:
      tags: [sandboxes]
      summary: List sandbox services
      description: Lists canonical sandbox services derived from `services` or the legacy `public_gateway` compatibility projection.
      security:
        - bearerAuth: []
      x-upstream-service: manager
      x-upstream-path: /api/v1/sandboxes/{id}/services
      parameters:
        - $ref: "#/components/parameters/SandboxID"
      responses:
        "200":
          description: Sandbox services
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessSandboxServicesResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
    put:
      tags: [sandboxes]
      summary: Replace sandbox services
      description: Replaces the canonical sandbox services. Public gateway remains available as a compatibility projection.
      security:
        - bearerAuth: []
      x-upstream-service: manager
      x-upstream-path: /api/v1/sandboxes/{id}/services
      parameters:
        - $ref: "#/components/parameters/SandboxID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SandboxServicesUpdateRequest"
      responses:
        "200":
          description: Sandbox services updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessSandboxServicesResponse"
        "400":
          description: Invalid sandbox services
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
...
  /api/v1/templates:
    get:
      tags: [templates]
      summary: List templates
      security:
        - bearerAuth: []
      x-upstream-service: manager
      x-upstream-path: /api/v1/templates
      responses:
        "200":
          description: Templates list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessTemplateListResponse"
    post:
      tags: [templates]
      summary: Create template
      security:
        - bearerAuth: []
      x-upstream-service: manager
      x-upstream-path: /api/v1/templates
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TemplateCreateRequest"
      responses:
        "201":
          description: Template created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessTemplateResponse"
...
  /api/v1/credential-sources:
    get:
      tags: [credential-sources]
      summary: List credential sources
      security:
        - bearerAuth: []
      x-upstream-service: manager
      x-upstream-path: /api/v1/credential-sources
      responses:
        "200":
          description: Credential sources
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessCredentialSourceListResponse"
    post:
      tags: [credential-sources]
      summary: Create credential source
      security:
        - bearerAuth: []
      x-upstream-service: manager
      x-upstream-path: /api/v1/credential-sources
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CredentialSourceWriteRequest"
      responses:
        "201":
          description: Credential source created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessCredentialSourceResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    SandboxID:
      name: id
      in: path
      required: true
      schema:
        type: string
...
  schemas:
...
    ErrorEnvelope:
      type: object
      properties:
        success:
          type: boolean
          enum: [false]
        error:
          $ref: "#/components/schemas/Error"
      required: [success, error]
...
    ClaimRequest:
      type: object
      properties:
        template:
          type: string
        config:
          $ref: "#/components/schemas/SandboxConfig"
        mounts:
          type: array
          items:
            $ref: "#/components/schemas/ClaimMountRequest"
      required: []
...
    SuccessClaimResponse:
      allOf:
        - $ref: "#/components/schemas/SuccessEnvelope"
        - type: object
          properties:
            data:
              $ref: "#/components/schemas/ClaimResponse"
...
    SuccessSandboxListResponse:
      allOf:
        - $ref: "#/components/schemas/SuccessEnvelope"
        - type: object
          properties:
            data:
              type: object
              properties:
                sandboxes:
                  type: array
                  items:
                    $ref: "#/components/schemas/SandboxSummary"
                count:
                  type: integer
                  description: Total matching sandboxes
                has_more:
                  type: boolean
                  description: More results available
              required:
                - sandboxes
                - count
                - has_more
...